CRPL-2902: Committing to preserve most the extensive protection and logging version 	-  The fully isolated version seems to always work with ganache 		-  Background downloading implicitely of multiple solc compiler versions at the same time is prevented by explicitely allowing only one codepath enter to the solc downloading at a time. 		-  The "lock" is released strictly outside of the  solcSnapshot's callback function. (To remain on the safe side, since no idea, what external process execution, threading, etc. interference can happen and what is the exact lifecycle of the downloaded solcSnapshot process) 		-  Output ports - with either the result or an error - are invoked from outside of the solcSnapshot callback scope.
/Compilation precedes the deployment on the data flow. Extensive lifecycle logging reveals that the async downloading and compilation (the normal usage pattern) by multiple solcSnapshots can overlap in time.
This seems to cause seemingly unrelated errors downstream in later nodes, e.g. in deployment./

Since the error with ganache is non-deterministic, telling for sure which technique prevented the error in deployment is hard.
This commit seems to solve it but may not be necessary to apply both modifications, also CheckReady may be replaced by less implicit techniques (however a simple setTimeOut(...,0) from inside the callback do not seem to solve alone the problem by just breaking out from the callback's callstack.

The commit serves as a candidate checkpoint to return to, in case simplifications turn out to be insuficient.